+++ package.json
@@ -15,7 +15,7 @@
         "react-dom": "^17.0.1",
         "react-router-dom": "^5.2.0",
         "react-scripts": "4.0.1",
-        "tailwindcss": "npm:@tailwindcss/postcss7-compat",
+        "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.3",
         "web-vitals": "^0.2.4"
     },
     "scripts": {
